Add empty members list to repo config
Peribolos can manage org members using a "members" list in the yaml
config. I don't have any additional github users to use right now as
both users I have access to are already org admins and can't be in both
the 'admins' and 'members' lists, so just add this as an empty list for
now.